# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Intune Policy Auditor API",
    "description": "Read-only, deterministic analysis of exported Intune policies.",
    "version": "3.0.0-dev"
  },
  "paths": {
...
    "/api/v1/audits": {
      "post": {
        "tags": [
          "Audits"
        ],
        "summary": "Create Offline Audit",
        "operationId": "create_offline_audit_api_v1_audits_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_offline_audit_api_v1_audits_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_AuditResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audits/{audit_id}": {
      "get": {
        "tags": [
          "Audits"
        ],
        "summary": "Get Audit",
        "operationId": "get_audit_api_v1_audits__audit_id__get",
        "parameters": [
          {
            "name": "audit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Audit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_AuditResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssignmentRisk": {
        "properties": {
          "risk_id": {
            "type": "string",
            "title": "Risk Id"
          },
          "policy_id": {
            "type": "string",
            "title": "Policy Id"
          },
          "risk_type": {
            "type": "string",
            "title": "Risk Type"
          },
          "severity": {
            "$ref": "#/components/schemas/FindingSeverity"
          },
          "explanation": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Explanation"
          },
          "recommended_action": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Recommended Action"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "risk_id",
          "policy_id",
          "risk_type",
          "severity",
          "explanation",
          "recommended_action"
        ],
        "title": "AssignmentRisk"
      },
...
      "AuditConfiguration": {
        "properties": {
          "language": {
            "type": "string",
            "pattern": "^(de|en)$",
            "title": "Language",
            "default": "de"
          },
          "active_pack_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Active Pack Ids"
          },
          "exact_only": {
            "type": "boolean",
            "title": "Exact Only",
            "default": true
          },
          "history_enabled": {
            "type": "boolean",
            "title": "History Enabled",
            "default": false
          },
          "save_reports": {
            "type": "boolean",
            "title": "Save Reports",
            "default": false
          },
          "context": {
            "$ref": "#/components/schemas/AuditContext"
          },
          "organization_requirements": {
            "items": {
              "$ref": "#/components/schemas/OrganizationRequirement"
            },
            "type": "array",
            "title": "Organization Requirements"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "active_pack_ids"
        ],
        "title": "AuditConfiguration"
      },
      "AuditContext": {
        "properties": {
          "os_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Os Version"
          },
          "edition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edition"
          },
          "architecture": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Architecture"
          },
          "enrollment_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enrollment Type"
          },
          "supervised": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supervised"
          },
          "granted_licenses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Granted Licenses"
          },
          "physical_device": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Physical Device"
          },
          "vdi": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vdi"
          },
          "shared_device": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shared Device"
          },
          "pilot_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pilot Context"
          },
          "environment_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Environment Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AuditContext"
      },
      "AuditMode": {
        "type": "string",
        "enum": [
          "offline",
          "tenant",
          "demo"
        ],
        "title": "AuditMode"
      },
...
      "AuditResult": {
        "properties": {
          "audit_id": {
            "type": "string",
            "title": "Audit Id"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "3.0"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "mode": {
            "$ref": "#/components/schemas/AuditMode"
          },
          "is_synthetic": {
            "type": "boolean",
            "title": "Is Synthetic"
          },
          "configuration": {
            "$ref": "#/components/schemas/AuditConfiguration"
          },
          "policies": {
            "items": {
              "$ref": "#/components/schemas/PolicyEvaluation"
            },
            "type": "array",
            "title": "Policies"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Findings"
          },
          "conflicts": {
            "items": {
              "$ref": "#/components/schemas/ConflictResult"
            },
            "type": "array",
            "title": "Conflicts"
          },
          "assignment_risks": {
            "items": {
              "$ref": "#/components/schemas/AssignmentRisk"
            },
            "type": "array",
            "title": "Assignment Risks"
          },
          "diagnostics": {
            "items": {
              "$ref": "#/components/schemas/ParserDiagnostic"
            },
            "type": "array",
            "title": "Diagnostics"
          },
          "coverage": {
            "$ref": "#/components/schemas/CoverageMetrics"
          },
          "decision": {
            "$ref": "#/components/schemas/DecisionResult"
          },
          "selected_pack_manifests": {
            "items": {
              "$ref": "#/components/schemas/KnowledgePackManifest"
            },
            "type": "array",
            "title": "Selected Pack Manifests"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "audit_id",
          "created_at",
          "mode",
          "is_synthetic",
          "configuration",
          "policies",
          "findings",
          "conflicts",
          "assignment_risks",
          "diagnostics",
          "coverage",
          "decision",
          "selected_pack_manifests"
        ],
        "title": "AuditResult"
      },
...
      "Body_create_offline_audit_api_v1_audits_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files",
            "description": "Intune JSON exports or safe ZIP archives"
          },
          "knowledge_pack_ids": {
            "type": "string",
            "title": "Knowledge Pack Ids",
            "default": "[\"synthetic.test-baseline\"]"
          },
          "language": {
            "type": "string",
            "pattern": "^(de|en)$",
            "title": "Language",
            "default": "de"
          },
          "deviations": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deviations"
          },
          "organization_requirements": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Requirements"
          },
          "audit_context": {
            "type": "string",
            "title": "Audit Context",
            "default": "{}"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_create_offline_audit_api_v1_audits_post"
      },
...
      "ConflictResult": {
        "properties": {
          "conflict_id": {
            "type": "string",
            "title": "Conflict Id"
          },
          "confidence": {
            "$ref": "#/components/schemas/ConflictConfidence"
          },
          "classification": {
            "type": "string",
            "title": "Classification"
          },
          "canonical_setting_id": {
            "type": "string",
            "title": "Canonical Setting Id"
          },
          "policy_a_id": {
            "type": "string",
            "title": "Policy A Id"
          },
          "policy_a_name": {
            "type": "string",
            "title": "Policy A Name"
          },
          "policy_a_value": {
            "$ref": "#/components/schemas/JsonValue"
          },
          "policy_b_id": {
            "type": "string",
            "title": "Policy B Id"
          },
          "policy_b_name": {
            "type": "string",
            "title": "Policy B Name"
          },
          "policy_b_value": {
            "$ref": "#/components/schemas/JsonValue"
          },
          "assignments_a": {
            "items": {
              "$ref": "#/components/schemas/AssignmentTarget"
            },
            "type": "array",
            "title": "Assignments A"
          },
          "assignments_b": {
            "items": {
              "$ref": "#/components/schemas/AssignmentTarget"
            },
            "type": "array",
            "title": "Assignments B"
          },
          "overlap_result": {
            "type": "string",
            "title": "Overlap Result"
          },
          "reasoning": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Reasoning"
          },
          "affected_device_count": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Affected Device Count"
          },
          "recommended_action": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Recommended Action"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "conflict_id",
          "confidence",
          "classification",
          "canonical_setting_id",
          "policy_a_id",
          "policy_a_name",
          "policy_a_value",
          "policy_b_id",
          "policy_b_name",
          "policy_b_value",
          "assignments_a",
          "assignments_b",
          "overlap_result",
          "reasoning",
          "recommended_action"
        ],
        "title": "ConflictResult"
      },
      "CoverageMetrics": {
        "properties": {
          "parser_coverage": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Parser Coverage"
          },
          "technical_extraction_confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Technical Extraction Confidence"
          },
          "exact_setting_coverage": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Exact Setting Coverage"
          },
          "exact_value_semantic_coverage": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Exact Value Semantic Coverage"
          },
          "microsoft_baseline_coverage": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Microsoft Baseline Coverage"
          },
          "documentation_coverage": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Documentation Coverage"
          },
          "assignment_analysis_coverage": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignment Analysis Coverage"
          },
          "conflict_analysis_coverage": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Conflict Analysis Coverage"
          },
          "runtime_evidence_coverage": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Evidence Coverage"
          },
          "evaluable_settings": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Evaluable Settings"
          },
          "not_evaluable_settings": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Not Evaluable Settings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "evaluable_settings",
          "not_evaluable_settings"
        ],
        "title": "CoverageMetrics"
      },
...
      "DecisionResult": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/DecisionStatus"
          },
          "title": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Title"
          },
          "reason": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Reason"
          },
          "blocking_conditions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blocking Conditions"
          },
          "nonblocking_warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Nonblocking Warnings"
          },
          "next_action": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Next Action"
          },
          "evidence_summary": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Evidence Summary"
          },
          "failed_gates": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Failed Gates"
          },
          "decision_trace": {
            "items": {
              "$ref": "#/components/schemas/DecisionTraceStep"
            },
            "type": "array",
            "title": "Decision Trace"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "title",
          "reason",
          "blocking_conditions",
          "nonblocking_warnings",
          "next_action",
          "evidence_summary",
          "failed_gates",
          "decision_trace"
        ],
        "title": "DecisionResult"
      },
...
      "Finding": {
        "properties": {
          "finding_id": {
            "type": "string",
            "title": "Finding Id"
          },
          "severity": {
            "$ref": "#/components/schemas/FindingSeverity"
          },
          "finding_type": {
            "$ref": "#/components/schemas/FindingType"
          },
          "policy_id": {
            "type": "string",
            "title": "Policy Id"
          },
          "policy_name": {
            "type": "string",
            "title": "Policy Name"
          },
          "setting_id": {
            "type": "string",
            "title": "Setting Id"
          },
          "setting_name": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Setting Name"
          },
          "configured_value": {
            "$ref": "#/components/schemas/JsonValue"
          },
          "microsoft_value": {
            "$ref": "#/components/schemas/JsonValue"
          },
          "alignment": {
            "$ref": "#/components/schemas/AlignmentStatus"
          },
          "raw_alignment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AlignmentStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Title"
          },
          "explanation": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Explanation"
          },
          "why_it_matters": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Why It Matters"
          },
          "user_impact": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "User Impact"
          },
          "business_impact": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Business Impact"
          },
          "recommended_action": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Recommended Action"
          },
          "pilot_suggestion": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Pilot Suggestion"
          },
          "rollback_suggestion": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Rollback Suggestion"
          },
          "evidence_confidence": {
            "$ref": "#/components/schemas/EvidenceConfidence"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceSource"
            },
            "type": "array",
            "title": "Evidence"
          },
          "baseline_product": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Product"
          },
          "baseline_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Version"
          },
          "verification_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Date"
          },
          "decision_trace": {
            "items": {
              "$ref": "#/components/schemas/DecisionTraceStep"
            },
            "type": "array",
            "title": "Decision Trace"
          },
          "assignments": {
            "items": {
              "$ref": "#/components/schemas/AssignmentTarget"
            },
            "type": "array",
            "title": "Assignments"
          },
          "affected_device_count": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Affected Device Count"
          },
          "deviation_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deviation State"
          },
          "accepted_deviation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AcceptedDeviation"
              },
              {
                "type": "null"
              }
            ]
          },
          "applicability": {
            "$ref": "#/components/schemas/ApplicabilityState"
          },
          "runtime_evidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeEvidence"
              },
              {
                "type": "null"
              }
            ]
          },
          "not_evaluable_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Not Evaluable Reasons"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "finding_id",
          "severity",
          "finding_type",
          "policy_id",
          "policy_name",
          "setting_id",
          "setting_name",
          "configured_value",
          "alignment",
          "title",
          "explanation",
          "why_it_matters",
          "user_impact",
          "business_impact",
          "recommended_action",
          "pilot_suggestion",
          "rollback_suggestion",
          "evidence_confidence",
          "evidence",
          "decision_trace",
          "assignments",
          "applicability"
        ],
        "title": "Finding"
      },
      "FindingSeverity": {
        "type": "string",
        "enum": [
          "critical",
          "high",
          "medium",
          "low",
          "information",
          "none"
        ],
        "title": "FindingSeverity"
      },
      "FindingType": {
        "type": "string",
        "enum": [
          "baseline_deviation",
          "confirmed_conflict",
          "probable_conflict",
          "possible_conflict",
          "assignment_risk",
          "broad_assignment",
          "missing_assignment",
          "unsupported_setting",
          "deprecated_setting",
          "runtime_error",
          "runtime_conflict",
          "expired_deviation",
          "insufficient_evidence"
        ],
        "title": "FindingType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JsonValue": {},
...
      "ParserDiagnostic": {
        "properties": {
          "source_file": {
            "type": "string",
            "title": "Source File"
          },
          "policy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy"
          },
          "json_path": {
            "type": "string",
            "title": "Json Path"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "severity": {
            "$ref": "#/components/schemas/DiagnosticSeverity"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "suggested_action": {
            "type": "string",
            "title": "Suggested Action"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_file",
          "json_path",
          "code",
          "severity",
          "reason",
          "suggested_action"
        ],
        "title": "ParserDiagnostic"
      },
...
      "PolicyEvaluation": {
        "properties": {
          "policy": {
            "$ref": "#/components/schemas/PolicyDocument"
          },
          "settings": {
            "items": {
              "$ref": "#/components/schemas/SettingEvaluation"
            },
            "type": "array",
            "title": "Settings"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Findings"
          },
          "assignment_risks": {
            "items": {
              "$ref": "#/components/schemas/AssignmentRisk"
            },
            "type": "array",
            "title": "Assignment Risks"
          },
          "runtime_evidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeEvidence"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "policy",
          "settings",
          "findings"
        ],
        "title": "PolicyEvaluation"
      },
...
      "ResponseEnvelope_AuditResult_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AuditResult"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ResponseEnvelope[AuditResult]"
      },
...
      "ResponseMeta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "ResponseMeta"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}